+++ src/CrossPointState.cpp
@@ -97,9 +97,11 @@ bool CrossPointState::loadFromBinaryFile() {
   if (version >= 5) {
     serialization::readPod(inputFile, pendingBookmarkSpine);
     serialization::readPod(inputFile, pendingBookmarkProgress);
+    pendingBookmarkParagraphIndex = UINT16_MAX;
   } else {
     pendingBookmarkSpine = UINT16_MAX;
     pendingBookmarkProgress = -1.0f;
+    pendingBookmarkParagraphIndex = UINT16_MAX;
   }
 
   return true;